Refactor and add navigation item
