+++ packages/edit-site/src/components/left-sidebar/navigation-panel/navigation-entity-items.js
